revert settings and fix users views
