+++ src/views/RegularUsers.tsx
@@ -249,11 +249,10 @@ export const RegularUsers = () => {
 
     return (
         <>
-            <Container style={{paddingTop: "40px"}}>
+            <Container style={{padding: "0px"}}>
                 <Row>
                     <Col span={24}>
-                        <Title level={4}>Users</Title>
-                        <Paragraph>A list of all users{(window.location.hostname == "app.netbird.io") ? ". RegularUsers with an email from the same private organization domain will automatically join when they sign in for the first time." : ""}</Paragraph>
+                        <Paragraph>Manage the users and their permissions.{(window.location.hostname == "app.netbird.io") ? "Same-domain email users are added automatically on first sign-in." : ""}</Paragraph>
                         <Space direction="vertical" size="large" style={{display: 'flex'}}>
                             <Row gutter={[16, 24]}>
                                 <Col xs={24} sm={24} md={8} lg={8} xl={8} xxl={8} span={8}>
